requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
Django==4.0
django-ckeditor==6.2.0
django-filter==21.1
django-import-export==2.7.0
django-js-asset==1.2.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
import-export==0.2.67.dev6
Pillow==8.4.0
requests==2.27.1
virtualenv==20.13.0
2captcha-python==1.1.2
altgraph==0.17.2
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
attrs==21.4.0
autopep8==1.7.0
beautifulsoup4==4.10.0
botocore==1.29.7
bs4==0.0.1
capmonster-python==2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dill==0.3.5.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-filter==21.1
django-import-export==2.7.0
django-js-asset==1.2.2
django-pwa==1.0.10
django-rest-framework==0.1.0
django-social-share==2.3.0
django-storages==1.13.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
filelock==3.7.1
future==0.18.2
h11==0.13.0
idna==3.3
import-export==0.2.67.dev6
importlib-metadata==4.12.0
isort==5.10.1
jmespath==1.0.1
lazy-object-proxy==1.7.1
MarkupPy==1.14
mccabe==0.7.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pefile==2022.5.30
Pillow==8.4.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
pwinput==1.0.2
pybrowsers==0.5.0
pycodestyle==2.9.1
pycparser==2.21
pyinstaller==5.2
